Avoid fastacmd failure if no reannos for this organism

+++ cgi/downloadReanno.cgi
@@ -40,26 +40,25 @@ my $reanno = $dbh->selectall_arrayref("SELECT * from Reannotation JOIN Gene USIN
 my $filename = $orgId ? "reanno_$orgId.tsv" : "reanno.tsv";
 
 # Fetch the sequences
-my $procId = $$;
-my $timestamp = int (gettimeofday * 1000);
-my $pre = Utils::tmp_dir() . "/" . $procId . $timestamp;
-my $listFile = "$pre.list";
-my $faaFile = "$pre.faa";
-open(LIST, ">", $listFile) || die "Cannot write to $listFile";
-foreach my $row (@$reanno) {
-  print LIST join(":", $row->{orgId}, $row->{locusId}) . "\n";
+my $aaseqs = {}; # lcl|orgId:locusId => sequence
+if (@$reanno > 0) {
+  my $procId = $$;
+  my $timestamp = int (gettimeofday * 1000);
+  my $pre = Utils::tmp_dir() . "/" . $procId . $timestamp;
+  my $listFile = "$pre.list";
+  my $faaFile = "$pre.faa";
+  open(LIST, ">", $listFile) || die "Cannot write to $listFile";
+  foreach my $row (@$reanno) {
+    print LIST join(":", $row->{orgId}, $row->{locusId}) . "\n";
+  }
+  close(LIST) || die "Error writing to $listFile";
+  my $fastacmd = '../bin/blast/fastacmd';
+  system($fastacmd, '-d', $myDB, '-i', $listFile, '-o', $faaFile) == 0
+    || die "Error running $fastacmd -- $!";
+  $aaseqs = FEBA_Utils::ReadFasta($faaFile);
+  unlink($listFile);
+  unlink($faaFile);
 }
-close(LIST) || die "Error writing to $listFile";
-
-my $fastacmd = '../bin/blast/fastacmd';
-system($fastacmd, '-d', $myDB, '-i', $listFile, '-o', $faaFile) == 0
-  || die "Error running $fastacmd -- $!";
-
-# lcl|orgId:locusId => sequence
-my $aaseqs = FEBA_Utils::ReadFasta($faaFile);
-
-unlink($listFile);
-unlink($faaFile);
 
 print "Content-Type:text/tab-separated-values\n";
 print "Content-Disposition: attachment; filename=$filename\n\n";
